cloud-run-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.16.0
gunicorn==21.2.0
python-dotenv==1.0.0
